requirements.txt
# Core data stack
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Templating
jinja2>=3.1.0

# Config
pyyaml>=6.0

# Testing
pytest>=7.4.0

# Data versioning (install separately — needs network)
# dvc>=3.0.0
# dvc[s3]>=3.0.0     # for AWS S3 remote
# dvc[gdrive]>=3.0.0 # for Google Drive remote

# psycopg2-binary>=2.9.0
